+++ packages/styles/bases/menu.css.ts
@@ -8,21 +8,29 @@ export const menuItemStyle = style({
 
 export const menuItemVariantStyle = styleVariants({
   default: {
-    color: THEME.color.gray[600],
+    color: THEME.color.gray[700],
+    backgroundColor: THEME.color.white[800],
     cursor: "pointer",
   },
   hover: {
     color: THEME.color.green[800],
+    backgroundColor: THEME.color.gray[200],
     cursor: "pointer",
   },
   active: {
     color: THEME.color.green[800],
-    backgroundColor: THEME.color.white[800],
-    fontWeight: 700,
+    backgroundColor: THEME.color.green[300],
     cursor: "pointer",
   },
   disabled: {
-    color: THEME.color.gray[500],
+    color: THEME.color.gray[700],
+    backgroundColor: THEME.color.white[800],
+    opacity: 0.5,
     cursor: "not-allowed",
+  },
+  selected: {
+    color: THEME.color.green[800],
+    backgroundColor: THEME.color.white[800],
+    fontWeight: 700,
   },
 });