+++ packages/wiz-ui-react/src/components/base/popup/utils/index.ts
@@ -1,2 +1,2 @@
-export * from "./placement";
-export { wrapDirection, wrapOutOfBound } from "./wrap";
+export * from "./popup-position";
+export * from "./rect";