+++ packages/wiz-ui-react/src/components/icons/index.ts
@@ -16,9 +16,11 @@ import { WizIChat } from "./chat";
 import { WizICheck } from "./check";
 import { WizIChevronLeft } from "./chevron-left";
 import { WizIChevronRight } from "./chevron-right";
+import { WizICircle } from "./circle";
 import { WizICircleCheck } from "./circle-check";
 import { WizIClose } from "./close";
 import { WizIContentCopy } from "./content-copy";
+import { WizIContract } from "./contract";
 import { WizIDashboard } from "./dashboard";
 import { WizIDescription } from "./description";
 import { WizIDownload } from "./download";
@@ -31,6 +33,7 @@ import { WizIHelp } from "./help";
 import { WizIHistory } from "./history";
 import { WizIHome } from "./home";
 import { WizIInfo } from "./info";
+import { WizILocationCity } from "./location-city";
 import { WizIManageAccounts } from "./manage-accounts";
 import { WizIMenu } from "./menu";
 import { WizIMonitoring } from "./monitoring";
@@ -44,6 +47,8 @@ import { WizIRemove } from "./remove";
 import { WizISchedule } from "./schedule";
 import { WizISearch } from "./search";
 import { WizISend } from "./send";
+import { WizISettings } from "./settings";
+import { WizISmartphone } from "./smartphone";
 import { WizIStars } from "./stars";
 import { WizIStoreFront } from "./store-front";
 import { WizITab } from "./tab";
@@ -71,8 +76,10 @@ export type TIcon =
   | typeof WizIChevronLeft
   | typeof WizIChevronRight
   | typeof WizICircleCheck
+  | typeof WizICircle
   | typeof WizIClose
   | typeof WizIContentCopy
+  | typeof WizIContract
   | typeof WizIDashboard
   | typeof WizIDescription
   | typeof WizIDownload
@@ -85,6 +92,7 @@ export type TIcon =
   | typeof WizIHistory
   | typeof WizIHome
   | typeof WizIInfo
+  | typeof WizILocationCity
   | typeof WizIManageAccounts
   | typeof WizIMenu
   | typeof WizIMonitoring
@@ -98,6 +106,8 @@ export type TIcon =
   | typeof WizISchedule
   | typeof WizISearch
   | typeof WizISend
+  | typeof WizISettings
+  | typeof WizISmartphone
   | typeof WizIStars
   | typeof WizIStoreFront
   | typeof WizITab
@@ -125,8 +135,10 @@ export {
   WizIChevronLeft,
   WizIChevronRight,
   WizICircleCheck,
+  WizICircle,
   WizIClose,
   WizIContentCopy,
+  WizIContract,
   WizIDashboard,
   WizIDescription,
   WizIDownload,
@@ -139,6 +151,7 @@ export {
   WizIHistory,
   WizIHome,
   WizIInfo,
+  WizILocationCity,
   WizIManageAccounts,
   WizIMenu,
   WizIMonitoring,
@@ -152,6 +165,8 @@ export {
   WizISchedule,
   WizISearch,
   WizISend,
+  WizISettings,
+  WizISmartphone,
   WizIStars,
   WizIStoreFront,
   WizITab,