+++ packages/wiz-ui/src/components/base/header/header.stories.ts
@@ -1,9 +1,7 @@
 import { StoryFn } from "@storybook/vue";
 import { THEME, SPACING_ACCESSORS } from "@wizleap-inc/wiz-ui-constants";
-import { provide } from "vue";
 
 import { WizTextButton } from "@/components";
-import { globalKey, useGlobalProvider } from "@/hooks/use-global-provider";
 
 import { WizHeader } from ".";
 
@@ -22,13 +20,13 @@ export default {
     sticky: {
       control: { type: "boolean" },
     },
+    onToggle: {
+      action: "click",
+    },
   },
   decorators: [
     (story: StoryFn) => ({
       components: { story },
-      setup() {
-        provide(globalKey, useGlobalProvider());
-      },
       template: `<story />`,
     }),
   ],
@@ -38,7 +36,7 @@ const Template: StoryFn = (_, { argTypes }) => ({
   props: Object.keys(argTypes),
   components: { WizHeader },
   template: `
-    <wiz-header v-bind="$props"/>
+    <wiz-header v-bind="$props" @onToggle="onToggle('menu')" />
   `,
 });
 
@@ -48,7 +46,7 @@ const LeftSlotTemplate: StoryFn = (_, { argTypes }) => ({
   props: Object.keys(argTypes),
   components: { WizHeader, WizTextButton },
   template: `
-    <wiz-header v-bind="$props">
+    <wiz-header v-bind="$props" @onToggle="onToggle('menu')">
       <template #left>
         <wiz-text-button>Click</wiz-text-button>
         <wiz-text-button>Click</wiz-text-button>
@@ -68,7 +66,7 @@ const RightSlotTemplate: StoryFn = (_, { argTypes }) => ({
   props: Object.keys(argTypes),
   components: { WizHeader, WizTextButton },
   template: `
-    <wiz-header v-bind="$props">
+    <wiz-header v-bind="$props" @onToggle="onToggle('menu')">
       <template #right>
         <wiz-text-button>Click</wiz-text-button>
         <wiz-text-button>Click</wiz-text-button>
@@ -84,11 +82,12 @@ RightSlotWithGap.args = {
   gapRight: "sm",
 };
 
-export const Fixed: StoryFn = () => ({
+export const Fixed: StoryFn = (_, { argTypes }) => ({
+  props: Object.keys(argTypes),
   components: { WizHeader, WizTextButton },
   template: `
     <div style="height: 200vh">
-      <wiz-header sticky>
+      <wiz-header @onToggle="onToggle('menu')" sticky>
         <template #left>
           <wiz-text-button>Click</wiz-text-button>
           <wiz-text-button>Click</wiz-text-button>